--- a/apps/linear_algebra/Makefile
+++ b/apps/linear_algebra/Makefile
@@ -276,32 +276,32 @@ $(KERNEL_DIR)/halide_dger_impl.o $(KERNEL_DIR)/halide_dger_impl.h: $(KERNEL_DIR)
 
 $(KERNEL_DIR)/halide_sgemm_notrans.o $(KERNEL_DIR)/halide_sgemm_notrans.h: $(KERNEL_DIR)/blas_l3.generator
 	$(LD_PATH_SETUP) $< -g sgemm -f halide_sgemm_notrans -o $(KERNEL_DIR) -e $(EMIT_OPTIONS) \
-	target=$(HL_TARGET_NR) parallel=false vectorize=true transpose_A=false transpose_B=false
+	target=$(HL_TARGET_NR) transpose_A=false transpose_B=false
 
 $(KERNEL_DIR)/halide_dgemm_notrans.o $(KERNEL_DIR)/halide_dgemm_notrans.h: $(KERNEL_DIR)/blas_l3.generator
 	$(LD_PATH_SETUP) $< -g dgemm -f halide_dgemm_notrans -o $(KERNEL_DIR) -e $(EMIT_OPTIONS) \
-	target=$(HL_TARGET_NR) parallel=false vectorize=true transpose_A=false transpose_B=false
+	target=$(HL_TARGET_NR) transpose_A=false transpose_B=false
 
 $(KERNEL_DIR)/halide_sgemm_transA.o $(KERNEL_DIR)/halide_sgemm_transA.h: $(KERNEL_DIR)/blas_l3.generator
 	$(LD_PATH_SETUP) $< -g sgemm -f halide_sgemm_transA -o $(KERNEL_DIR) -e $(EMIT_OPTIONS) \
-	target=$(HL_TARGET_NR) parallel=false vectorize=true transpose_A=true transpose_B=false
+	target=$(HL_TARGET_NR) transpose_A=true transpose_B=false
 
 $(KERNEL_DIR)/halide_dgemm_transA.o $(KERNEL_DIR)/halide_dgemm_transA.h: $(KERNEL_DIR)/blas_l3.generator
 	$(LD_PATH_SETUP) $< -g dgemm -f halide_dgemm_transA -o $(KERNEL_DIR) -e $(EMIT_OPTIONS) \
-	target=$(HL_TARGET_NR) parallel=false vectorize=true transpose_A=true transpose_B=false
+	target=$(HL_TARGET_NR) transpose_A=true transpose_B=false
 
 $(KERNEL_DIR)/halide_sgemm_transB.o $(KERNEL_DIR)/halide_sgemm_transB.h: $(KERNEL_DIR)/blas_l3.generator
 	$(LD_PATH_SETUP) $< -g sgemm -f halide_sgemm_transB -o $(KERNEL_DIR) -e $(EMIT_OPTIONS) \
-	target=$(HL_TARGET_NR) parallel=false vectorize=true transpose_A=false transpose_B=true
+	target=$(HL_TARGET_NR) transpose_A=false transpose_B=true
 
 $(KERNEL_DIR)/halide_dgemm_transB.o $(KERNEL_DIR)/halide_dgemm_transB.h: $(KERNEL_DIR)/blas_l3.generator
 	$(LD_PATH_SETUP) $< -g dgemm -f halide_dgemm_transB -o $(KERNEL_DIR) -e $(EMIT_OPTIONS) \
-	target=$(HL_TARGET_NR) parallel=false vectorize=true transpose_A=false transpose_B=true
+	target=$(HL_TARGET_NR) transpose_A=false transpose_B=true
 
 $(KERNEL_DIR)/halide_sgemm_transAB.o $(KERNEL_DIR)/halide_sgemm_transAB.h: $(KERNEL_DIR)/blas_l3.generator
 	$(LD_PATH_SETUP) $< -g sgemm -f halide_sgemm_transAB -o $(KERNEL_DIR) -e $(EMIT_OPTIONS) \
-	target=$(HL_TARGET_NR) parallel=false vectorize=true transpose_A=true transpose_B=true
+	target=$(HL_TARGET_NR) transpose_A=true transpose_B=true
 
 $(KERNEL_DIR)/halide_dgemm_transAB.o $(KERNEL_DIR)/halide_dgemm_transAB.h: $(KERNEL_DIR)/blas_l3.generator
 	$(LD_PATH_SETUP) $< -g dgemm -f halide_dgemm_transAB -o $(KERNEL_DIR) -e $(EMIT_OPTIONS) \
-	target=$(HL_TARGET_NR) parallel=false vectorize=true transpose_A=true transpose_B=true
+	target=$(HL_TARGET_NR) transpose_A=true transpose_B=true