… apps/linear_algebra/benchmarks/clock.h
… apps/linear_algebra/benchmarks/halide_benchmarks.cpp
… apps/linear_algebra/benchmarks/macros.h
… apps/linear_algebra/tests/test_halide_blas.cpp
… src/RemoveDeadAllocations.cpp
… src/Simplify.cpp