Used roughly the same memory order as before (z-order over output tiles) but improved utilization of vector registers in the inner loop, which required tiled storage of one of the inputs.

### Sandy Bridge perf

Master (last three columns are matrix size, microseconds, and gflops):

  Halide              sgemm_notrans      32            4.799329             7.46771
  Halide              sgemm_notrans      64           21.910316             12.5252
  Halide              sgemm_notrans     128          150.002623             14.3084
  Halide              sgemm_notrans     288          405.989838             59.4515
  Halide              sgemm_notrans     544         1593.410873             101.591
  Halide              sgemm_notrans    1056        10979.063416             107.562
  Halide              sgemm_notrans    2080        90380.749512             99.7103

This branch:

  Halide              sgemm_notrans      32            1.881409             19.0496
  Halide              sgemm_notrans      64           13.252832             20.7074
  Halide              sgemm_notrans     128           45.612109             47.0556
  Halide              sgemm_notrans     288          345.321875             69.8962
  Halide              sgemm_notrans     544         1220.325000             132.651
  Halide              sgemm_notrans    1056         8285.300000             142.533
  Halide              sgemm_notrans    2080        65423.500000             137.747

OpenBLAS is the fastest of the reference blasses on my machine:

OpenBLAS              sgemm_notrans      32            2.930957             12.2281
OpenBLAS              sgemm_notrans      64           48.324609             5.67893
OpenBLAS              sgemm_notrans     128           74.773047             28.7042
OpenBLAS              sgemm_notrans     288          388.653125             62.1035
OpenBLAS              sgemm_notrans     544         1432.725000             112.985
OpenBLAS              sgemm_notrans    1056         8929.350000             132.253
OpenBLAS              sgemm_notrans    2080        68494.750000             131.571

### Haswell perf

Old:

  Halide              sgemm_notrans      32            4.029369             8.89469
  Halide              sgemm_notrans      64           15.015531             18.2765
  Halide              sgemm_notrans     128           50.655842             42.3703
  Halide              sgemm_notrans     288          274.934387             87.7908
  Halide              sgemm_notrans     544         1592.437744             101.654
  Halide              sgemm_notrans    1056        12555.395508             94.0575
  Halide              sgemm_notrans    2080       105421.643066             85.4843

New:

  Halide              sgemm_notrans      32            1.059271             33.8346
  Halide              sgemm_notrans      64            7.121973             38.5331
  Halide              sgemm_notrans     128           31.174219             68.8487
  Halide              sgemm_notrans     288          170.959375             141.184
  Halide              sgemm_notrans     544          935.531250             173.032
  Halide              sgemm_notrans    1056         7084.100000             166.701
  Halide              sgemm_notrans    2080        69218.100000             130.196

OpenBLAS:

OpenBLAS              sgemm_notrans      32            2.268323             15.8002
OpenBLAS              sgemm_notrans      64           10.424512             26.3256
OpenBLAS              sgemm_notrans     128           26.461328              81.111
OpenBLAS              sgemm_notrans     288          199.682812             120.875
OpenBLAS              sgemm_notrans     544         1033.031250             156.701
OpenBLAS              sgemm_notrans    1056         7044.350000             167.642
OpenBLAS              sgemm_notrans    2080        53774.900000             167.585

On this Haswell machine we fall out of the cache at 2080, because we don't block over the sum axis.